testing/web-platform/tests/2dcontext/drawing-images-to-the-canvas/drawimage_html_image_5_srgb.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="dest" height="100" width="100"></canvas>

<script>
var sourceWidth = 100;
var sourceHeight = 100;
var smoothingEnabled = false;
var destCanvas = document.getElementById('dest');
var sourceImg = document.createElement('img');
sourceImg.src = '../2x2-srgb.png'
sourceImg.width = sourceWidth;
sourceImg.height = sourceHeight;
var destCtx = destCanvas.getContext('2d');

function checkPixel(x, y, expected, epsilon) {
    var actual = destCtx.getImageData(x, y, 1, 1).data;
    assert_true(actual.length === expected.length);
    for (var i=0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], epsilon);
}

var pixelTests = [
    ['Pixel Top Left Canvas Background', 0, 0, [255, 0, 0, 255], 1],
    ['Pixel Top Right Canvas Background', 0, 99, [255, 0, 0, 255], 1],
    ['Pixel Bottom Left Canvas Background', 99, 0, [255, 0, 0, 255], 1],
    ['Pixel Bottom Right Canvas Background', 99, 99, [255, 0, 0, 255], 1],

    ['Pixel Surrounding Top Left Canvas Background', 49, 49, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Top Right Canvas Background', 49, 82, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Bottom Left Canvas Background', 99, 49, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Bottom Right Canvas Background', 99, 82, [255, 0, 0, 255], 1],

    ['Pixel Top Left Image', 50, 50, [253, 140, 245, 255], 1],
    ['Pixel Top Right Image', 50, 81, [41, 122, 115, 255], 1],
    ['Pixel Bottom Left Image', 99, 50, [41, 122, 115, 255], 1],
    ['Pixel Bottom Right Image', 99, 81, [253, 140, 245, 255], 1],
];

function runTest() {
    destCtx.fillStyle = "#FF0000";
    destCtx.fillRect(0, 0,  destCanvas.width, destCanvas.height);
    destCtx.imageSmoothingEnabled = smoothingEnabled;
    // The source canvas will copied to the 50,50 position of the destination canvas
    // over an area of 64x32 pixels
    // The copied image will be distorted along the x axis
    destCtx.drawImage(sourceImg, 50, 50, 64, 32);
    generate_tests(checkPixel, pixelTests);
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Draw 100x100 image to 100x100 canvas at 50,50 with resize to 64x32.');
</script>
